added que

+++ resources/views/playlist.blade.php
@@ -4,12 +4,14 @@
 
 
 <div class="container">
+    <h1>Playlist: {{$playlist->name}} Duration: {{$duration}}</h1>
     <table>
         <tr>
             <th>Song:</th>
-            <th>Created at:</th>
-            <th>Updated at:</th>
+            <th>Artist</th>
+            <th>Duration</th>
             <th>Play</th>
+            <th>Details</th>
             <th>Delete song from playlist</th>
 
         </tr>
@@ -17,12 +19,14 @@
         @for ($i=0; $i<count($song); $i++)
             <tr>
                 <td><h3>{{$song[$i]->name}}</h3></td>
-                <td>{{$song[$i]->created_at}}</td>
-                <td>{{$song[$i]->updated_at}}</td>
-                <td><a>Play</a></td>
-                <td><a href="/playlist/delete/{{$song[$i]->id}}">Delete</a></td>
+                <td>{{$song[$i]->artist}}</td>
+                <td>{{$song[$i]->duration}}</td>
+                <td><a href="/que/add/{{$song[$i]->id}}">Play</a></td>
+                <td><a href="/song/{{$song[$i]->id}}" >Details</a></td>
+                <td><a href="/playlist/delete/{{$song[$i]->id}}" onclick="return confirm('Are you sure you want to delete this item?')">Delete</a></td>
             </tr>
         @endfor
     </table>
 </div>
 @endsection
+
